last factory output

+++ task2/src/main/java/by/shift/task2/core/model/impl/Rectangle.java
@@ -13,8 +13,8 @@ public class Rectangle implements Figure {
     double square;
     double diagonal;
 
-    int sideB;
-    int sideA;
+    double sideB;
+    double sideA;
 
     @Override
     public void calculate(FileData fileData) {
